pyproject.toml
[tool.poetry]
name = "autoapply"
version = "0.1.0"
description = "An innovative solution for automating job applications using LinkedIn profile data"
authors = ["Ericson Willians <ericsonwillians@protonmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.13.1"
playwright = "^1.41.1"
python-dotenv = "^1.0.0"
transformers = "^4.36.2"
pdfplumber = "^0.10.3"
jsonschema = "^4.21.1"
typer = "^0.9.0"
rich = "^13.7.0"
pydantic = "^2.5.3"
python-logstash = "^0.4.8"
structlog = "^24.1.0"
torch = "^2.5.1"
pydantic-settings = "^2.7.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"
pylint = "^3.0.3"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
autoapply = "app.main:app"
